requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
arrow==1.4.0
beautifulsoup4==4.14.3
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
coverage==7.13.5
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.135.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hypothesis==6.152.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
joblib==1.5.3
libcst==1.8.6
linkify-it-py==2.1.0
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mutmut==3.5.0
metatrader5==5.0.5735
numpy==2.2.6
openai==2.30.0
packaging==26.0
pandas==2.3.3
platformdirs==4.9.6
pluggy==1.6.0
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.2
pytest-cov==7.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
rich==15.0.0
scikit-learn==1.7.2
scipy==1.15.3
setproctitle==1.3.7
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
starlette==1.0.0
textual==8.2.3
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.4.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uc-micro-py==2.0.0
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
xgboost==3.2.0
